Move the VersionFieldExtensions class to the Extentsions subfolder
