+++ src/GitVersion.Core/Extensions/VersionFieldExtensions.cs
@@ -1,4 +1,4 @@
-namespace GitVersion;
+namespace GitVersion.Extensions;
 
 internal static class VersionFieldExtensions
 {
