+++ src/GitVersion.Core/VersionCalculation/TrunkBased/EnrichIncrement.cs
@@ -1,4 +1,5 @@
 using GitVersion.Configuration;
+using GitVersion.Extensions;
 
 namespace GitVersion.VersionCalculation.TrunkBased;
